move serialization to another module
